Refresh user permissions when access token gets updated
